Issue #277: additional enhancements to the work done in r600 to enable the details update to use multiple data providers in order (filling empty fields when possible). This change also eliminates the need for some nodes put into "hybrid" scraper scripts (like mymovies.it).
* added overrides to custom type StringList to allow basic comparison
* fixed datatype REAL returning it's default value as double instead of a single (float)
* added ProtectExistingValuesFromCopy() to control the outcome of CopyUpdatableValues()
* added new advanced setting DataProviderRequestLimit which let the user control how many data providers the imported should try/use when searching/updating movie details.
* updated several scraper scripts to implement best practise (use site_id not details_url) and optimized them. (still in the process of reviewing other available scripts)

+++ trunk/Cornerstone/Database/DBField.cs
@@ -117,7 +117,7 @@ namespace Cornerstone.Database {
                             return int.Parse(attribute.Default);
                     case DBDataType.REAL:
                         if (attribute.Default == "")
-                            return 0.0;
+                            return (float)0.0;
                         else
                             return float.Parse(attribute.Default);
                     case DBDataType.BOOL:
